+++ crd/src/lombok/com/nextbreakpoint/flink/k8s/crd/V1BootstrapSpec.java
@@ -2,6 +2,7 @@ package com.nextbreakpoint.flink.k8s.crd;
 
 import com.google.gson.annotations.SerializedName;
 import io.kubernetes.client.openapi.models.V1ResourceRequirements;
+import io.kubernetes.client.openapi.models.V1SecurityContext;
 import lombok.Builder;
 import lombok.Data;
 import lombok.AllArgsConstructor;
@@ -30,4 +31,6 @@ public class V1BootstrapSpec {
     private String serviceAccount;
     @SerializedName("resources")
     private V1ResourceRequirements resources;
+    @SerializedName("securityContext")
+    private V1SecurityContext securityContext;
 }