+++ src/main/kotlin/com/nextbreakpoint/flink/k8s/supervisor/task/JobOnStopped.kt
@@ -54,6 +54,13 @@ class JobOnStopped : Task<JobManager>() {
             return
         }
 
+        if (manager.isJobRunning()) {
+            manager.onJobStarted()
+            return
+        }
+
+        manager.updateJobStatus()
+
         if (!manager.shouldRestartJob()) {
             return
         }